feat: Add PR template
