+++ src/createStore.js
@@ -1,0 +1,9 @@
+import {createStore,applyMiddleware} from 'redux'
+import RootReducer from './reducers/RootReducer';
+import ReduxThunk from 'redux-thunk'
+
+export const middlewares =[ReduxThunk]
+
+export const createStoreWithMiddleware = applyMiddleware(...middlewares)(createStore)
+
+export const store = createStoreWithMiddleware(RootReducer)
