+++ src/reducers/RootReducer.js
@@ -1,0 +1,6 @@
+import {combineReducers} from 'redux';
+import success from './SuccessReducer';
+
+export default combineReducers({
+    success
+})
