--- teesco-frontend/src/store/createStore.js
@@ -1,9 +1,0 @@
-import {createStore,applyMiddleware} from 'redux'
-import RootReducer from './reducers/rootReducer';
-import Logger from 'redux-logger'
-import ReduxThunk from 'redux-thunk'
-import {composeWithDevTools} from 'redux-devtools-extension'
-
-export const middlewares =[ReduxThunk,Logger]
-
-export const store = createStore(RootReducer,composeWithDevTools(applyMiddleware(...middlewares)))
