--- teesco-frontend/src/store/reducers/rootReducer.js
@@ -1,6 +1,0 @@
-import {combineReducers} from 'redux';
-import success from './successReducer';
-
-export default combineReducers({
-    success
-})
